# backend/internal/server/openapi.yaml
openapi: 3.0.3
info:
  title: Go-Svelte Starter Kit API
  description: |
    Public REST API secured by API keys.
    Pass your key in the `Authorization: Bearer sk_<key>` header.
  version: 1.0.0
servers:
  - url: /api/v1
    description: Versioned public API

security:
  - ApiKeyAuth: []

components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: "API key issued from /profile/api-keys. Format: Bearer sk_<key>"

  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string

    Profile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        display_name:
          type: string
        avatar_url:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        email_verified_at:
          type: string
          format: date-time
          nullable: true
        two_fa_enabled:
          type: boolean
        created_at:
          type: string
          format: date-time

    UpdateProfileRequest:
      type: object
      properties:
        display_name:
          type: string
          minLength: 1
          maxLength: 100
        bio:
          type: string
          maxLength: 500

    Notification:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [info, success, warning, alert]
        title:
          type: string
        body:
          type: string
          nullable: true
        link:
          type: string
          nullable: true
        read_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time

    NotificationList:
      type: object
      properties:
        notifications:
          type: array
          items:
            $ref: '#/components/schemas/Notification'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        display_name:
          type: string
        avatar_url:
          type: string
          nullable: true
        email_verified_at:
          type: string
          format: date-time
          nullable: true
        two_fa_enabled:
          type: boolean
        roles:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time

    Webhook:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
        events:
          type: array
          items:
            type: string
        active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    CreateWebhookRequest:
      type: object
      required:
        - url
        - events
      properties:
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
          minItems: 1

paths:
  /me:
    get:
      summary: Get authenticated user profile
      operationId: getProfile
      security:
        - ApiKeyAuth: [read:profile]
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Profile'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      summary: Update authenticated user profile
      operationId: updateProfile
      security:
        - ApiKeyAuth: [write:profile]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Profile'
        '401':
          description: Unauthorized

  /notifications:
    get:
      summary: List notifications
      operationId: listNotifications
      security:
        - ApiKeyAuth: [read:notifications]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: unread
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Notification list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/NotificationList'

  /notifications/{id}:
    patch:
      summary: Mark a notification as read
      operationId: markNotificationRead
      security:
        - ApiKeyAuth: [write:notifications]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Notification marked as read
        '404':
          description: Not found

  /users:
    get:
      summary: List users (admin scope required)
      operationId: listUsers
      security:
        - ApiKeyAuth: [read:users]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: User list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      users:
                        type: array
                        items:
                          $ref: '#/components/schemas/User'
                      total:
                        type: integer

  /webhooks:
    get:
      summary: List webhooks
      operationId: listWebhooks
      security:
        - ApiKeyAuth: [read:webhooks]
      responses:
        '200':
          description: Webhook list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Webhook'
    post:
      summary: Create a webhook
      operationId: createWebhook
      security:
        - ApiKeyAuth: [write:webhooks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWebhookRequest'
      responses:
        '201':
          description: Webhook created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Webhook'

  /webhooks/{id}:
    delete:
      summary: Delete a webhook
      operationId: deleteWebhook
      security:
        - ApiKeyAuth: [write:webhooks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deleted
        '404':
          description: Not found